requirements.txt
# Core Testing Framework
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
pytest-ordering==0.6

# Web UI Automation - Playwright
playwright==1.49.1
pytest-playwright==0.6.2

# Mobile Automation - Appium
Appium-Python-Client==4.1.0
selenium==4.27.1

# API Testing
requests==2.32.3
httpx==0.28.1
jsonschema==4.23.0

# Reporting - Allure
allure-pytest==2.13.5
allure-python-commons==2.13.5

# Configuration & Utilities
pydantic==2.10.3
PyYAML==6.0.2

# Logging & Debugging
loguru==0.7.3
rich==13.9.4

# Data Generation
Faker==33.1.0

# Async Support
anyio==4.7.0
